% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vocal_props.R
\name{get_vocal_props}
\alias{get_vocal_props}
\title{Get Frog Call Properties}
\usage{
get_vocal_props(path = "examples/")
}
\arguments{
\item{path}{Folder directory containing the calls.}
}
\value{
A data frame containing the extracted call properties.
}
\description{
This custom function automates the extraction of 15 commonly used temporal and spectral call properties for analyzing frog calls.
Cleaned and standardized call samples must be placed in one folder, and the call format should be in .wav.
}
\examples{
call.dir <- system.file("extdata", "calls", package = "VocalProps")

prop.df <- get_vocal_props(path = call.dir)

head(prop.df)

## write results to a csv 
#write.csv(prop.df, "frog_call_prop.csv", row.names = FALSE)

}
\references{
Köhler J, Jansen M, Rodríguez A, Kok PJR, Toledo LF, Emmrich M, Glaw F, Haddad CFB, Rödel M-O, Vences M (2017). The use of bioacoustics in anuran taxonomy: theory, terminology, methods and recommendations for best practice. Zootaxa 4251: 1–124. https://doi.org/10.11646/zootaxa.4251.1.1

Xie J, Towsey M, Zhang J, Roe P (2018) Frog call classification: a survey. Artificial Intelligence Review, 49: 375–391.

Xie J, Towsey M, Zhang J, Roe P (2020) Investigation of Acoustic and Visual Features for Frog Call Classification. Journal of Signal Processing Systems, 92: 23–36.

Prasad VK, Chuang M-F, Das A, Ramesh K, Yi Y, Dinesh KP, Borzée A (2022) Coexisting good neighbours: acoustic and calling microhabitat niche partitioning in two elusive syntopic species of balloon frogs, Uperodon systoma and U. globulosus (Anura: Microhylidae) and potential of individual vocal signatures. BMC Zoology, 7: 27.
}
